Use regular classes for Configuration in README
Since `Configuration` classes don't represent an ADT, we should really
be suggesting they be modelled with a case class hierarchy.

By using a regular class over a case class, you don't just gain the
ability to provide default values, but you are _forced_ to, which is a
good thing.

For values that don't have a rational default, you can still provide a
default that would fail its validation (e.g. `null` when the value is
`@NotNull`).