Fix exceptions in timed blocks not stopping timer
When using a `timed` block (i.e. through `TimerWrapper#timed`), if the
user function `f` throws an exception, the timer would never be stopped.

To fix this, we simply need to wrap the evaluation in a try/finally
block to ensure the timer is always stopped, even when an exception is
thrown.