Add test utils for Application, MySQL, Liquibase
Each utility class enables a test (using ScalaTest) to execute within
the context of the utility. If multiple utilities are defined, they will
be started and stopped in the order defined (just like Dropwizard's
`Managed`), with the tests running within the context of all of them.

See `ScalaApplicationSpecIT` in `core/test` for an example of their use.

Note: the MySQL and Liquibase utilities provide an optional dependency.
To use these, you will need to add an explicit dependency on either
`mysql-mxj` or `dropwizard-migrations` to the `Test` configuration of
your project.