Fix mapping result sets with NULL in some rows
When a column is nullable, if there is at least one column after the
nullable column, the nullable column contains a `NULL` value,
that `NULL` value will be incorrectly applied to the _next_ column.

This is caused by `rs.wasNull`, which looks at the _previous_ value
retrieved, not the current one.